Move triggers out of migrate directory
