+++ packages/data/setup.py
@@ -40,13 +40,13 @@ PACKAGES = [
     'galaxy.model.dataset_collections',
     'galaxy.model.dataset_collections.types',
     'galaxy.model.migrate',
-    'galaxy.model.migrate.triggers',
     'galaxy.model.migrate.versions',
     'galaxy.model.orm',
     'galaxy.model.store',
     'galaxy.model.tool_shed_install',
     'galaxy.model.tool_shed_install.migrate',
     'galaxy.model.tool_shed_install.migrate.versions',
+    'galaxy.model.triggers',
     'galaxy.model.unittest_utils',
     'galaxy.model.view',
     'galaxy.quota',
